Change IntSet::len() to return u64 instead of usize.
A u32 containing intset can have up to u32::MAX + 1 elements in it which can overflow a usize when usize is a u32.

+++ fuzz/fuzz_targets/fuzz_int_set.rs
@@ -14,7 +14,7 @@ use int_set_op_processor::OperationSet;
 use int_set_op_processor::SmallEvenInt;
 use int_set_op_processor::SmallInt;
 
-const OPERATION_COUNT: usize = 7_500;
+const OPERATION_COUNT: u64 = 7_500;
 
 fuzz_target!(|data: &[u8]| {
     let Some(mode_byte) = data.first() else {
